+++ CHANGELOG.md
@@ -1,14 +1,15 @@
 # Change Log
 
 ## Unreleased
+- persons: country_short_blue, country_short_white を入れた方が良くないか？
 
 ## 0.2.0-SNAPSHOT
 - src/fights/db.clj create tables. `persons`, `contests` and `competitions`.
 - test/fights/db_test.clj
 - src/fights/competitor.clj
 - src/fights/core.clj
-- can insert-person
-
+- (insert-person-id id_competition)
+- (insert-pserons year)
 
 ## 0.1.0 - 2022-08-12
 - project (re)started.